# [Consensus] Switch to generic binary search