Let's not support a custom `bsearch(3)` implementation.  As a side benefit new generic-based version is slightly faster:
```
$ benchstat old.txt new.txt                                                                                           
name               old time/op    new time/op    delta
LeaderSelection-8     520ms ± 0%     508ms ± 1%  -2.27%  (p=0.000 n=9+10)

name               old alloc/op   new alloc/op   delta
LeaderSelection-8    30.0MB ± 0%    30.0MB ± 0%    ~     (p=0.797 n=9+10)

name               old allocs/op  new allocs/op  delta
LeaderSelection-8      99.0 ± 0%      99.0 ± 0%    ~     (all equal)
```