@AlexHentschel 2022-11-10
How about the following modifications:
1. we include the following additional documentation [here](https://github.com/onflow/flow-go/blob/cf8364cec57bb56561b57de35949ec953323aed9/consensus/hotstuff/committees/leader/leader_selection.go#L164)
    ```golang
    // CAUTION: we must find the element with the **smallest index**, whose value is
    // equal or larger than `randomness`. For example, consider the slice 
    //   weightSums = [0, 0, 2, 5, 5, 8, 16, 16]
    // and assume we have randomness = 5. For correctness of the leader selection, we
    // require that index `3` is returned. However, at the time of writing this code, 
    // the _implementation_ of `slices.BinarySearch` satisfies our requirement to find
    // the smallest index. But the API documentation does not guarantee this property. 
    // In other words, we are relying on non-documented, internal implementation details
    // of `slices.BinarySearch`. We verify that this critical property is retained by
    // future versions of golang through the unit test Test_SmallestIndexReturned.
    ```
2. include the following fuzzy test:
    ```golang
    // Test_SmallestIndexReturned verifies that `slices.BinarySearch` satisfies the following property:
    //
    //	`slices.BinarySearch` returns the **smallest index**, whose value is equal or
    //	 larger than `target`
    //
    // This property is critical for the correctness of our leader selection. For example, consider the slice
    //
    //	weightSums = [0, 0, 2, 5, 5, 8, 16, 16]
    //
    // and assume we have randomness = 5. For correctness of the leader selection, we
    // require that index `3` is returned. However, at the time of writing this code,
    // the _implementation_ of `slices.BinarySearch` satisfies our requirement to find
    // the smallest index. But the API documentation does not guarantee this property.
    // In other words, we are relying on non-documented, internal implementation details
    // of `slices.BinarySearch`. We verify that this critical property is retained by
    // future versions of golang through this unit test.
    func Test_SmallestIndexReturned(t *testing.T) {
    	for sliceLength := 1; sliceLength <= 123; sliceLength++ {
    		for repetition := 10; repetition > 0; repetition-- {
    			firstIndex := rand.Intn(sliceLength)
    			weightSums := make([]uint64, sliceLength)
    			for k := 0; k < firstIndex; k++ {
    				weightSums[k] = 7
    			}
    			for k := firstIndex; k < sliceLength; k++ {
    				weightSums[k] = 13
    			}

    			i, found := slices.BinarySearch(weightSums, 13)
    			require.True(t, found)
    			require.Equal(t, firstIndex, i)
    		}
    	}
    }
    ```